scripts/requirements.txt
requests>=2.31.0
web3>=6.0.0
eth-utils>=2.2.0
eth-keys>=0.4.0
click>=8.1.0
cryptography>=3.4.8
pycryptodome>=3.15.0
mnemonic>=0.20